fix(student): restrict to delete student data if they have issued book

+++ backend/src/student/student.service.ts
@@ -6,12 +6,13 @@ import { UpdateStudentDto } from './dto/update-student.dto';
 import { Repository, FindOptionsWhere, Like } from 'typeorm';
 import { Student } from './entities/student.entity';
 import { SettingsService } from 'src/settings/settings.service';
+import { Issue } from 'src/issues/entities/issue.entity';
 @Injectable()
 export class StudentService {
   constructor(
     @InjectRepository(Student)
     private readonly studentRepository: Repository<Student>,
-
+    @InjectRepository(Issue) private readonly issueRepo: Repository<Issue>,
     private readonly settingService: SettingsService,
   ) {}
 
@@ -71,7 +72,18 @@ export class StudentService {
     return this.studentRepository.update({ id }, updateStudentDto);
   }
 
-  remove(id: string) {
+  async remove(id: string) {
+    const issue = await this.issueRepo.findOne({
+      where: {
+        studentId: id,
+        returned: false,
+      },
+    });
+    if (issue) {
+      throw new BadRequestException(
+        'This student has issued book. This student cannot be deleted',
+      );
+    }
     return this.studentRepository.delete(id);
   }
 }
